fix: 상품 추가/수정 페이지 수정

+++ src/main/java/com/spring_boots/spring_boots/item/dto/ResponseItemDto.java
@@ -14,7 +14,7 @@ public class ResponseItemDto {
     private Long id;
     private String itemName;
     private Long categoryId;
-    private Integer itemPrice;
+    private Long itemPrice;
     private String itemDescription;
     private String itemMaker;
     private List<String> itemColor;
